rename: 상수 폴더명 변경 (const -> constants)
